# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Retail Shelf Assistant API",
    "description": "Voice-enabled retail assistant for product location and information queries",
    "version": "1.0.0",
    "contact": {
      "name": "Retail Assistant Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/ask": {
      "post": {
        "summary": "Process text query",
        "description": "Process a text-based product query and return location or information response",
        "tags": ["Queries"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {"$ref": "#/components/schemas/ProductLocationResponse"},
                    {"$ref": "#/components/schemas/ProductInfoResponse"}
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/ask-voice": {
      "post": {
        "summary": "Process voice query",
        "description": "Process an audio file containing a product query",
        "tags": ["Queries"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file (WAV, MP3, M4A, FLAC)"
                  },
                  "session_id": {
                    "type": "string",
                    "description": "Optional session identifier"
                  }
                },
                "required": ["audio_file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {"$ref": "#/components/schemas/ProductLocationResponse"},
                    {"$ref": "#/components/schemas/ProductInfoResponse"}
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Check service health and component status",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HealthResponse"}
              }
            }
          },
          "503": {
            "description": "Service is unhealthy",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductMatch": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "description": "Unique product identifier"
          },
          "product_name": {
            "type": "string",
            "description": "Full product name"
          },
          "brand": {
            "type": "string",
            "description": "Product brand"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "aisle": {
            "type": "string",
            "description": "Store aisle"
          },
          "bay": {
            "type": "string",
            "description": "Bay number"
          },
          "shelf": {
            "type": "string",
            "description": "Shelf level"
          },
          "confidence": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "Match confidence score"
          }
        },
        "required": ["product_id", "product_name", "brand", "category", "aisle", "bay", "shelf", "confidence"]
      },
      "ProductLocationResponse": {
        "type": "object",
        "properties": {
          "query_type": {
            "type": "string",
            "default": "location",
            "description": "Query type identifier"
          },
          "normalized_product": {
            "type": "string",
            "description": "Normalized product name from query"
          },
          "matches": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ProductMatch"},
            "description": "List of matching products"
          },
          "disambiguation_needed": {
            "type": "boolean",
            "description": "Whether user needs to clarify product choice"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes or instructions"
          }
        },
        "required": ["query_type", "normalized_product", "matches", "disambiguation_needed"]
      },
      "ProductInfoResponse": {
        "type": "object",
        "properties": {
          "query_type": {
            "type": "string",
            "default": "information",
            "description": "Query type identifier"
          },
          "normalized_product": {
            "type": "string",
            "description": "Normalized product name from query"
          },
          "answer": {
            "type": "string",
            "description": "LLM-generated answer about the product"
          },
          "caveats": {
            "type": "string",
            "description": "Important disclaimers or limitations"
          },
          "confidence": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "Response confidence score"
          }
        },
        "required": ["query_type", "normalized_product", "answer", "confidence"]
      },
      "TextQueryRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "User's text query"
          },
          "session_id": {
            "type": "string",
            "description": "Optional session identifier"
          }
        },
        "required": ["query"]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Service status"
          },
          "version": {
            "type": "string",
            "description": "Application version"
          },
          "components": {
            "type": "object",
            "description": "Component health status"
          }
        },
        "required": ["status", "version", "components"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "error_code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": ["error", "error_code"]
      }
    }
  },
  "tags": [
    {
      "name": "Queries",
      "description": "Product query endpoints"
    },
    {
      "name": "System",
      "description": "System health and status"
    }
  ]
}